+++ R/mp_send.R
@@ -21,7 +21,7 @@
 #'
 #' To see event parameters, create custom fields in your GA4 account first, to see them in your reports 24hrs after you send them in with this function via `Custom definitions > Create custom dimensions` - `dimension name` will be how it looks like in the reports, `event parameter` will be the parameter you have sent in with the event.
 #'
-#' `user_id` can be used for [cross-platform analysis](https://support.google.com/analytics/answer/9213390)
+#' `user_id` can be used for cross-platform analysis.
 #'
 #' `timestamp_micros` should only be set to record events that happened in the past. This value can be overridden via user_property or event timestamps. Events can be backdated up to 48 hours. Note microseconds, not milliseconds.
 #'
@@ -214,4 +214,12 @@ mp_connection <- function(measurement_id,
 
 is.mp_connection <- function(x){
   inherits(x, "mp_connection")
+}
+
+#' @export
+print.mp_connection <- function(x, ...){
+  cat("\n==GA4 MP Connection\n")
+  cat("Measurement ID:  ", x$measurement_id)
+  cat("\nAPI Secret:       <OBSCURED in .Renviron MP_SECRET>")
+  cat("\nEndpoint:        ", x$endpoint)
 }