+++ NEWS.md
@@ -1,3 +1,9 @@
+# measurementProtocol v0.1.1
+
+* Add Docker image with plumber installed so it can be used with PubSub, available at `gcr.io/gcer-public/measurementProtocol`
+* Add `mp_parse_json` that lets you more easily parse generic JSON requests into Measurement Protocol hits, and `mp_parse_gtm` to parse out specifically GA4 events sent to GTM
+* Remove unsuppressable start up message "No consent file found" (CRAN check)
+
 # measurementProtocol v0.1.0
 
 * Initial release